--- src/main/java/org/example/loaders/Loader.java
@@ -1,7 +1,0 @@
-package org.example.loaders;
-
-import org.example.Inventory;
-
-public interface Loader<T> {
-    void loadFile();
-}
