+++ src/main/java/org/unlam/paradigmas/zeta/enums/Classification.java
@@ -1,4 +1,4 @@
-package org.example.enums;
+package org.unlam.paradigmas.zeta.enums;
 
 public enum Classification {
     FIRE,
