+++ src/main/java/org/unlam/paradigmas/zeta/loaders/Loader.java
@@ -1,0 +1,5 @@
+package org.unlam.paradigmas.zeta.loaders;
+
+public interface Loader<T> {
+    void loadFile();
+}
