+++ src/main/java/org/unlam/paradigmas/zeta/models/Library.java
@@ -1,4 +1,4 @@
-package org.example.models;
+package org.unlam.paradigmas.zeta.models;
 
 public record Library(String originTable, Recipe recipe) {
 }
