--- src/test/java/org/example/WorkerTest.java
@@ -1,4 +1,0 @@
-package org.example;
-
-public class WorkerTest {
-}
