+++ src/test/java/org/unlam/paradigmas/zeta/WorkerTest.java
@@ -1,0 +1,4 @@
+package org.unlam.paradigmas.zeta;
+
+public class WorkerTest {
+}
